refatoração listagem de posts
